# service-template/cli/cmd/template/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Swagger tools for {{PROVIDER}} integrator",
        "title": "{{SERVICE_TITLE}}",
        "contact": {
            "name": "@supplyTeam"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/{{BASE_PATH}}/search": {
            "post": {
                "description": "Endpoint for search process",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "prime-booking"
                ],
                "summary": "Search endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "default": "TIKETCOM",
                        "description": "Authentication header",
                        "name": "storeId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "WEB",
                        "description": "Authentication header",
                        "name": "channelId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "321321321",
                        "description": "Authentication header",
                        "name": "requestId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "guest",
                        "description": "Authentication header",
                        "name": "username",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "GATEWAY",
                        "description": "Authentication header",
                        "name": "serviceId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "resellerId",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "identity",
                        "in": "header"
                    },
                    {
                        "description": "Request Body",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/fare.IntegratorFareRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/fare.FlightIntegratorSearchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
        "/{{BASE_PATH}}/v2/search": {
            "post": {
                "description": "Endpoint for search process",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "prime-booking"
                ],
                "summary": "Search endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "default": "TIKETCOM",
                        "description": "Authentication header",
                        "name": "storeId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "WEB",
                        "description": "Authentication header",
                        "name": "channelId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "321321321",
                        "description": "Authentication header",
                        "name": "requestId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "guest",
                        "description": "Authentication header",
                        "name": "username",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "GATEWAY",
                        "description": "Authentication header",
                        "name": "serviceId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "resellerId",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "identity",
                        "in": "header"
                    },
                    {
                        "description": "Request Body",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/fare.IntegratorFareRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/fare.FlightIntegratorSearchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "enum.CabinType": {
            "type": "string",
            "enum": [
                "ECONOMY",
                "BUSINESS",
                "FIRST",
                "PREMIER",
                "PREMIUM_ECONOMY",
                "PREMIUM_BUSINESS",
                "PREMIUM_FIRST"
            ],
            "x-enum-varnames": [
                "ECONOMY",
                "BUSINESS",
                "FIRST",
                "PREMIER",
                "PREMIUM_ECONOMY",
                "PREMIUM_BUSINESS",
                "PREMIUM_FIRST"
            ]
        },
...
        "enum.FareRequestSource": {
            "type": "string",
            "enum": [
                "FD_RM",
                "DYNAMIC_SUPPLIER_MAPPING",
                "DEFAULT_RULE",
                "SAM"
            ],
            "x-enum-varnames": [
                "FD_RM",
                "DYNAMIC_SUPPLIER_MAPPING",
                "DEFAULT_RULE",
                "SAM"
            ]
        },
...
        "fare.Account": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "fare.BrandedFare": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "fare.Corporate": {
            "type": "object",
            "properties": {
                "airlines": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "corporateCodes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "corporateId": {
                    "type": "integer"
                }
            }
        },
        "fare.Currency": {
            "type": "object",
            "properties": {
                "departureCurrency": {
                    "type": "string"
                },
                "returnCurrency": {
                    "type": "string"
                }
            }
        },
        "fare.Fare": {
            "type": "object",
            "properties": {
                "brandedFare": {
                    "$ref": "#/definitions/fare.BrandedFare"
                },
                "distributionType": {
                    "type": "string"
                },
                "fareClassDetails": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fare.FareClassDetail"
                    }
                },
                "fareClasses": {
                    "type": "string"
                },
                "flexiFare": {
                    "type": "boolean"
                },
                "inclusiveAddons": {
                    "$ref": "#/definitions/fare.InclusiveAddons"
                },
                "originalAdultFare": {
                    "$ref": "#/definitions/model.FareDetail"
                },
                "originalChildFare": {
                    "$ref": "#/definitions/model.FareDetail"
                },
                "originalCurrency": {
                    "type": "string"
                },
                "originalInfantFare": {
                    "$ref": "#/definitions/model.FareDetail"
                },
                "originalTotalFare": {
                    "$ref": "#/definitions/model.FareDetail"
                },
                "otherFares": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fare.OtherFare"
                    }
                },
                "privateFare": {
                    "type": "boolean"
                },
                "supplierId": {
                    "type": "string"
                }
            }
        },
        "fare.FareClassDetail": {
            "type": "object",
            "properties": {
                "Destination": {
                    "type": "string"
                },
                "fareClass": {
                    "type": "string"
                },
                "origin": {
                    "type": "string"
                }
            }
        },
        "fare.FareSegment": {
            "type": "object",
            "properties": {
                "airlineCode": {
                    "type": "string"
                },
                "arrival": {
                    "type": "string"
                },
                "departure": {
                    "type": "string"
                },
                "fareSegmentTypes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fare.FareSegmentType"
                    }
                },
                "flightNumber": {
                    "type": "string"
                }
            }
        },
        "fare.FareSegmentDetail": {
            "type": "object",
            "properties": {
                "agt": {
                    "type": "number"
                },
                "airportTax": {
                    "type": "number"
                },
                "baseFare": {
                    "type": "number"
                },
                "commission": {
                    "type": "number"
                },
                "fareClass": {
                    "type": "string"
                },
                "flyThru": {
                    "type": "number"
                },
                "gst": {
                    "type": "number"
                },
                "iwjr": {
                    "type": "number"
                },
                "nta": {
                    "type": "number"
                },
                "psc": {
                    "type": "number"
                },
                "seatAvailable": {
                    "type": "integer"
                },
                "supplierDiscount": {
                    "type": "number"
                },
                "tax": {
                    "type": "number"
                },
                "total": {
                    "type": "number"
                },
                "unidentified": {
                    "type": "number"
                }
            }
        },
        "fare.FareSegmentType": {
            "type": "object",
            "properties": {
                "originalAdultFare": {
                    "$ref": "#/definitions/fare.FareSegmentDetail"
                },
                "originalChildFare": {
                    "$ref": "#/definitions/fare.FareSegmentDetail"
                },
                "originalCurrency": {
                    "type": "string"
                },
                "originalInfantFare": {
                    "$ref": "#/definitions/fare.FareSegmentDetail"
                },
                "originalTotalFare": {
                    "$ref": "#/definitions/fare.FareSegmentDetail"
                },
                "supplierId": {
                    "type": "string"
                }
            }
        },
        "fare.FlightIntegratorSearchResponse": {
            "type": "object",
            "properties": {
                "additionalData": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "corporateCodes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "departureFlights": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fare.SearchAvailabilityOW"
                    }
                },
                "integratorFareRequest": {
                    "$ref": "#/definitions/fare.IntegratorFareRequest"
                },
                "multiCities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fare.SearchMulticity"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "returnFlights": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fare.SearchAvailabilityOW"
                    }
                },
                "roundTripCombines": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fare.SearchCombine"
                    }
                },
                "roundTrips": {
                    "$ref": "#/definitions/fare.SearchAvailabilityRT"
                },
                "totalPage": {
                    "type": "integer"
                },
                "totalSize": {
                    "type": "integer"
                }
            }
        },
        "fare.InclusiveAddons": {
            "type": "object",
            "properties": {
                "departureScheduleAddonsList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fare.ScheduleInclusiveAddons"
                    }
                },
                "returnScheduleAddonsList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fare.ScheduleInclusiveAddons"
                    }
                }
            }
        },
        "fare.IntegratorFareRequest": {
            "type": "object",
            "properties": {
                "additionalData": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "adult": {
                    "type": "integer"
                },
                "blacklistMap": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "cabinClass": {
                    "$ref": "#/definitions/enum.CabinType"
                },
                "child": {
                    "type": "integer"
                },
                "corporate": {
                    "$ref": "#/definitions/fare.Corporate"
                },
                "currency": {
                    "$ref": "#/definitions/fare.Currency"
                },
                "departureDate": {
                    "type": "string"
                },
                "destination": {
                    "type": "string"
                },
                "distributionType": {
                    "type": "string"
                },
                "infant": {
                    "type": "integer"
                },
                "international": {
                    "type": "integer"
                },
                "origin": {
                    "type": "string"
                },
                "returnDate": {
                    "type": "string"
                },
                "routes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fare.Route"
                    }
                },
                "source": {
                    "$ref": "#/definitions/enum.FareRequestSource"
                },
                "supplierMappings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fare.SupplierMapping"
                    }
                },
                "supplierRequest": {
                    "$ref": "#/definitions/fare.SupplierRequest"
                },
                "tripTypes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_tiket_TIX-FLIGHT-COMMON-MODEL-GO_model_constant_enum.TripType"
                    }
                }
            }
        },
        "fare.OtherFare": {
            "type": "object",
            "properties": {
                "fareClasses": {
                    "type": "string"
                },
                "originalAdultFare": {
                    "$ref": "#/definitions/model.FareDetail"
                },
                "originalChildFare": {
                    "$ref": "#/definitions/model.FareDetail"
                },
                "originalInfantFare": {
                    "$ref": "#/definitions/model.FareDetail"
                },
                "originalTotalFare": {
                    "$ref": "#/definitions/model.FareDetail"
                }
            }
        },
        "fare.Route": {
            "type": "object",
            "properties": {
                "currency": {
                    "type": "string"
                },
                "departureDate": {
                    "type": "string"
                },
                "destination": {
                    "type": "string"
                },
                "origin": {
                    "type": "string"
                }
            }
        },
...
        "fare.SearchAvailability": {
            "type": "object",
            "properties": {
                "adult": {
                    "type": "integer"
                },
                "arrival": {
                    "$ref": "#/definitions/model.ScheduleDetail"
                },
                "child": {
                    "type": "integer"
                },
                "currency": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "departure": {
                    "$ref": "#/definitions/model.ScheduleDetail"
                },
                "destination": {
                    "type": "string"
                },
                "fareSellKey": {
                    "type": "string"
                },
                "flighSelect": {
                    "type": "string"
                },
                "flightSelect": {
                    "type": "string"
                },
                "infant": {
                    "type": "integer"
                },
                "journeySellKey": {
                    "type": "string"
                },
                "marketingAirline": {
                    "type": "string"
                },
                "origin": {
                    "type": "string"
                },
                "schedules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Schedule"
                    }
                },
                "seatsAvailable": {
                    "type": "integer"
                },
                "totalTransit": {
                    "type": "integer"
                },
                "upgradable": {
                    "type": "boolean"
                }
            }
        },
        "fare.SearchAvailabilityOW": {
            "type": "object",
            "properties": {
                "fareSegments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fare.FareSegment"
                    }
                },
                "fares": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fare.Fare"
                    }
                },
                "otherFares": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fare.Fare"
                    }
                },
                "searchAvailability": {
                    "$ref": "#/definitions/fare.SearchAvailability"
                }
            }
        },
        "fare.SearchAvailabilityRT": {
            "type": "object",
            "properties": {
                "departureFlights": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fare.SearchAvailabilityOW"
                    }
                },
                "returnFlights": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fare.SearchAvailabilityOW"
                    }
                }
            }
        },
        "fare.SearchCombine": {
            "type": "object",
            "properties": {
                "departureFlights": {
                    "$ref": "#/definitions/fare.SearchAvailability"
                },
                "fares": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fare.Fare"
                    }
                },
                "otherFares": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fare.Fare"
                    }
                },
                "returnFlights": {
                    "$ref": "#/definitions/fare.SearchAvailability"
                }
            }
        },
        "fare.SearchMulticity": {
            "type": "object",
            "properties": {
                "fares": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fare.Fare"
                    }
                },
                "otherFares": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fare.Fare"
                    }
                },
                "searchAvailabilities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fare.SearchAvailability"
                    }
                }
            }
        },
        "fare.SupplierMapping": {
            "type": "object",
            "properties": {
                "account": {
                    "$ref": "#/definitions/fare.Account"
                },
                "airlines": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "ignoreAirline": {
                    "type": "boolean"
                },
                "multiFare": {
                    "type": "boolean"
                }
            }
        },
        "fare.SupplierRequest": {
            "type": "object",
            "properties": {
                "accounts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fare.Account"
                    }
                },
                "airlines": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "github_com_tiket_TIX-FLIGHT-COMMON-MODEL-GO_model_constant_enum.TripType": {
            "type": "string",
            "enum": [
                "DEPARTURE",
                "RETURN",
                "ROUND_TRIP",
                "ROUND_TRIP_COMBINE",
                "MULTI_CITY"
            ],
            "x-enum-varnames": [
                "DEPARTURE",
                "RETURN",
                "ROUND_TRIP",
                "ROUND_TRIP_COMBINE",
                "MULTI_CITY"
            ]
        },
...
        "model.FareDetail": {
            "type": "object",
            "properties": {
                "agt": {
                    "type": "number"
                },
                "airportTax": {
                    "type": "number"
                },
                "base": {
                    "type": "number"
                },
                "breakdownTaxes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.TaxType"
                    }
                },
                "commission": {
                    "type": "number"
                },
                "fareBasisCodes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "fareClasses": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "flyThru": {
                    "type": "number"
                },
                "gst": {
                    "type": "number"
                },
                "iwjr": {
                    "type": "number"
                },
                "markup": {
                    "type": "number"
                },
                "nta": {
                    "type": "number"
                },
                "penaltyInfos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.PenaltyInfo"
                    }
                },
                "privateMarkup": {
                    "type": "number"
                },
                "psc": {
                    "type": "number"
                },
                "scheduleFareDetailInformations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ScheduleFareDetailInformation"
                    }
                },
                "seatAvailables": {
                    "type": "integer"
                },
                "supplierDiscount": {
                    "type": "number"
                },
                "tax": {
                    "type": "number"
                },
                "taxBreakDowns": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "total": {
                    "type": "number"
                },
                "unidentified": {
                    "type": "number"
                },
                "upselling": {
                    "type": "number"
                }
            }
        },
...
        "model.ScheduleDetail": {
            "type": "object",
            "properties": {
                "airportCode": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "terminal": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                },
                "timezone": {
                    "type": "number"
                }
            }
        },
...
    }
}